Renamed 'partials' everywhere to 'includes' Moved icons in static/ to static/icons
