+++ views/includes/svg.js
@@ -1,3 +1,3 @@
 module.exports = function(options) {
-  return require("fs").readFileSync(require("path").join(__dirname, '..', '..', 'public', options.svgBasename + '.svg'), 'utf8');
+  return require("fs").readFileSync(require("path").join(__dirname, '..', '..', 'static', 'icons', options.svgBasename + '.svg'), 'utf8');
 }
