+++ views/pages/manufacturers.js
@@ -3,7 +3,7 @@ module.exports = function(options) {
 
   options.title = 'Manufacturers - Open Fixture Library';
 
-  let str = require('../partials/header')(options);
+  let str = require('../includes/header')(options);
 
   str += '<h1>Manufacturers</h1>';
 
@@ -31,7 +31,7 @@ module.exports = function(options) {
     str += '</section>';
   }
 
-  str += require('../partials/footer')(options);
+  str += require('../includes/footer')(options);
 
   return str;
 };
