+++ views/pages/single_manufacturer.js
@@ -8,7 +8,7 @@ module.exports = function(options) {
   options.title = manufacturer.name + ' - Open Fixture Library';
 
   
-  let str = require('../partials/header')(options);
+  let str = require('../includes/header')(options);
 
   str += `<h1>${manufacturer.name} fixtures</h1>`;
 
@@ -31,7 +31,7 @@ module.exports = function(options) {
 
   str += '</ul>';
 
-  str += require('../partials/footer')(options);
+  str += require('../includes/footer')(options);
 
   return str;
 };
